FoobarMotorCopany/src/main/java/br/ifes/poo2/FoobarMotorCompany/cdp/Cidade.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package br.ifes.poo2.FoobarMotorCompany.cdp;

/**
 *
 * @author dolly
 */
public class Cidade {
    private String nome;
    private double latitude;
    private double longitude;

    public Cidade(String nome, double latitude, double longitude) {
        this.nome = nome;
        this.latitude = latitude;
        this.longitude = longitude;
    }

    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    public double getLatitude() {
        return latitude;
    }

    public void setLatitude(double latitude) {
        this.latitude = latitude;
    }

    public double getLongitude() {
        return longitude;
    }

    public void setLongitude(double longitude) {
        this.longitude = longitude;
    }
    
    public boolean equals (Object object){
        if(this == object){
            return true;
        }
        
        if(!(object instanceof Cidade)){
            return false;
        }
        
        Cidade cidade = (Cidade) object;
        return getNome().equals(cidade.getNome());
    }
    
    public String toString(){
        return getNome();
    }
    
}
